ImageMatting ImageExtractColor ImageDelColor
